% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{proj_install_deps}
\alias{proj_install_deps}
\title{Install dependencies}
\usage{
proj_install_deps(path = usethis::proj_get())
}
\arguments{
\item{path}{\code{character}, path to the project directory. If your current
working-directory is in the project, the default will do the right thing.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
Use to install the project's package dependencies from \code{DESCRIPTION}
using \code{\link[pak:local_install_deps]{pak::local_install_deps()}}.
}
\examples{
# not run because it produces side effects
if (FALSE) {
  proj_install_deps()
}
}
