% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{proj_use_workflow}
\alias{proj_use_workflow}
\title{Set up workflow directory}
\usage{
proj_use_workflow(path_proj = "analyses", git_ignore_data = TRUE)
}
\arguments{
\item{path_proj}{\code{character} path to the workflow directory,
relative to the project directory. Defaults to \code{"analyses"}.}

\item{git_ignore_data}{\code{logical} indicates to add \code{data} directory
contents to \code{.gitignore}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
Creates an \verb{analyses/} directory (or custom path) with a \verb{data/}
subdirectory, \code{.gitignore} rules, and a \code{README.md}.
}
\examples{
# not run because it creates side effects
\dontrun{
  proj_use_workflow()
}
}
