% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{use_qmd}
\alias{use_qmd}
\title{Use opinionated qmd template}
\usage{
use_qmd(
  name,
  path_proj = "analyses",
  open = rlang::is_interactive(),
  ignore = FALSE
)
}
\arguments{
\item{name}{\code{character} name of the workflow component.}

\item{path_proj}{\code{character} path to workflow directory,
relative to the project directory. Defaults to \code{"analyses"}.}

\item{open}{\code{logical} indicates to open the file for interactive editing.}

\item{ignore}{\code{logical} indicates to add this file to \code{.Rbuildignore}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
Use this function to create a new, templated Quarto file in a workflow
directory.
}
\details{
This is an opinionated system; it introduces restrictions to help keep you
"inside the boat":
\itemize{
\item All the Quarto files in a workflow are in the same directory; there
are no sub-directories with Quarto files.
\item Using \code{\link[here:i_am]{here::i_am()}} establishes the root of the workflow as the
directory that contains the Quarto file. In other words, when the
Quarto file is rendered, this directory becomes the \code{\link[here:here]{here::here()}} root.
\item It creates a dedicated data-directory for this file to write to, making
sure that this data-directory is empty at the start of the rendering. It
also provides an accessor function \code{path_target()} that you can use later
in the file to compose paths to this data-directory. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{write.csv(mtcars, path_target("mtcars.csv"))
}\if{html}{\out{</div>}}

It also provides an accessor function to the data directory itself, which
can be useful for reading data from "previous" files.

\if{html}{\out{<div class="sourceCode">}}\preformatted{fun_data <- read.csv(path_source("00-import", "fun_data.csv"))
}\if{html}{\out{</div>}}
}
}
\examples{
# not run because it creates side effects
\dontrun{
  # creates file `01-clean.qmd`
  use_qmd("01-clean")
}
}
