% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{use_qmd}
\alias{use_qmd}
\title{Use opinionated qmd template}
\usage{
use_qmd(
  name,
  path_proj = "analyses",
  open = rlang::is_interactive(),
  ignore = FALSE
)
}
\arguments{
\item{name}{\code{character} name of the workflow component.}

\item{path_proj}{\code{character} path to workflow directory,
relative to the project directory. Defaults to \code{"analyses"}.}

\item{open}{\code{logical} indicates to open the file for interactive editing.}

\item{ignore}{\code{logical} indicates to add this file to \code{.Rbuildignore}.}
}
\value{
Invisible \code{NULL}, called for side effects.
}
\description{
Use this function to create a new, templated Quarto file in a workflow
directory.
}
\examples{
# not run because it creates side effects
\dontrun{
  # creates file `01-clean.qmd`
  use_qmd("01-clean")
}
}
